Added JSON support for settings
I added JSON support. It now will run off of the settings inside the of
the settings.json file. I updated the checks.py but it isn't the best
solution, if any of you want to make it better please do so! Also please
adapt your cogs/modules to use it accordingly :)